// SmartBalance/Version 1.0.0/Code/Frontend/mersal-accounting-app/src/app/BusinessFeatures/Reports/Models/report-filter.model.ts
export class ReportSearch {
    dateTo: any;
    dateFrom: any;    
    accountChartId: any;
    currencyId: any;
    safeId: any;
}

export class VendorReportSearch {
    dateTo: any;
    dateFrom: any;
    vendorId: any;    
}

export class BankReportSearch {
    dateTo: any;
    dateFrom: any;
    bankId: any;    
}

export class CostCenterReportSearch {
    dateTo: any;
    dateFrom: any;
    costCenterId: any;    
}

export class BalanceSheetReportSearch {
    dateTo: any;
    dateFrom: any;    
    //accountChartId: any;
    level: any;
}

export class InventoryBalanceReport {
    dateTo: any;
    dateFrom: any;
    inventoryId: any;
}

export class FixedAssetDepreciationReportSearch {
    dateTo: any;
    dateFrom: any;
}

export class FixedAssetExcludedReportSearch {
    dateTo: any;
    dateFrom: any;
}

export class FixedAssetLocationReportSearch {
    dateTo: any;
    dateFrom: any;
    locationId: any;
}

export class FixedAssetInventoryReportSearch {
    dateTo: any;
    dateFrom: any;
    locationId: any;
}

export class FixedAssetLostReportSearch {
    dateTo: any;
    dateFrom: any;
    locationId: any;
}

export class DonatorCasesHistoryReport {
    dateTo: any;
    dateFrom: any;
    donatorId: any;
}

export class DonationCasesBalanceReport {
    dateTo: any;
    dateFrom: any;
    caseId: any;
}